# apps/backend/docs/swagger/swagger.yaml
basePath: /
definitions:
...
  http.CreateKeyRequest:
    properties:
      name:
        maxLength: 64
        type: string
    required:
    - name
    type: object
...
  http.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  http.LogoutRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  http.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  http.RegisterRequest:
    properties:
      email:
        type: string
      full_name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
    required:
    - email
    - full_name
    - password
    type: object
...
  http.TokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      token_type:
        type: string
      user:
        $ref: '#/definitions/http.UserResponse'
    type: object
  http.UserResponse:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      role:
        type: string
    type: object
  response.APIError:
    properties:
      code:
        type: string
      details:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  response.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.APIError'
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the Logify backend API server.
  title: Logify API
  version: "1.0"
paths:
...
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user credentials and return JWT tokens.
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully authenticated
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/http.TokenResponse'
              type: object
        "400":
          description: Invalid input format
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  type: string
              type: object
        "401":
          description: Invalid credentials
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  type: string
              type: object
      summary: Login a user
      tags:
      - auth
  /v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke the refresh token and clear user session.
      parameters:
      - description: Logout payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Logged out successfully
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Invalid request body
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  type: string
              type: object
      summary: Logout a user
      tags:
      - auth
  /v1/auth/refresh-token:
    post:
      consumes:
      - application/json
      description: Generate a new access token using a valid refresh token.
      parameters:
      - description: Refresh token payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/http.TokenResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  type: string
              type: object
      summary: Refresh token
      tags:
      - auth
  /v1/auth/register:
    post:
      consumes:
      - application/json
      description: Create a user account and its default-project atomically, then
        return JWT access and refresh tokens.
      parameters:
      - description: Registration request details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully registered and authenticated
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/http.TokenResponse'
              type: object
        "400":
          description: Invalid JSON or missing fields
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  type: string
              type: object
        "409":
          description: Email already in use
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  type: string
              type: object
      summary: Register a new user
      tags:
      - auth
...
  /v1/projects/{id}/api-keys:
    get:
      parameters:
      - description: Project UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: List project API keys
      tags:
      - API keys
    post:
      consumes:
      - application/json
      parameters:
      - description: Project UUID
        in: path
        name: id
        required: true
        type: string
      - description: Key name
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.CreateKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a project API key
      tags:
      - API keys
  /v1/projects/{id}/api-keys/{keyId}:
    delete:
      parameters:
      - description: Project UUID
        in: path
        name: id
        required: true
        type: string
      - description: API key UUID
        in: path
        name: keyId
        required: true
        type: string
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Revoke a project API key
      tags:
      - API keys
...
securityDefinitions:
  APIKeyAuth:
    description: Project API key, restricted to sending logs.
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Enter Bearer followed by your access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"